LoginViewDelegate
- Se crea LoginViewDelegate para manejar todas las views de la pantalla de login
