+++ app/src/main/res/values/colors.xml
@@ -7,11 +7,12 @@
     <color name="colorWhite">#FFFFFF</color>
     <color name="colorBlack">#000000</color>
     <color name="colorRed">#D50000</color>
+    <color name="colorTransparent">#80FFFFFF</color>
 
     <!--LOGIN-->
     <color name="loginBackground">#01579B</color>
     <color name="loginLogoTextColor">@color/colorWhite</color>
-    <color name="loginFormContainerColor">#80FFFFFF</color>
+    <color name="loginFormContainerColor">@color/colorTransparent</color>
     <color name="loginFormTitleTextColor">@color/colorBlack</color>
     <color name="loginFormButtonColor">@color/loginBackground</color>
     <color name="loginFormButtonTextColor">@color/colorWhite</color>
